+++ util.py
@@ -2,7 +2,7 @@ import tensorflow as tf
 import numpy as np
 
 
-def fc_layer(input_, input_size, output_size, activation=None, bias=True, scope=None):
+def fc_layer(input_, input_size, output_size, activation=None, bias=True, scope=None, dtype=tf.float32):
     '''
         fully connnected layer
         Args :
@@ -17,10 +17,10 @@ def fc_layer(input_, input_size, output_size, activation=None, bias=True, scope=
     '''
     with tf.variable_scope(scope or "fc"):
         w = tf.get_variable(name="weight", shape=[input_size, output_size],
-                            initializer=tf.contrib.layers.xavier_initializer())
+                            initializer=tf.contrib.layers.xavier_initializer(), dtype=dtype)
         output_ = tf.matmul(input_, w)
         if bias:
-            b = tf.get_variable(name="bias", shape=[output_size], initializer=tf.constant_initializer(0.001))
+            b = tf.get_variable(name="bias", shape=[output_size], initializer=tf.constant_initializer(0.001), dtype=dtype)
             output_ += b
         return output_ if activation is None else activation(output_)
